Move some functions into a new utils.py file to avoid circular dependencies
